Create IDEA2
My idea is to help autistic people, as we all know they are more sensitive to noises. Waking up with a loud alarm clock can be disturbing for them. That product will be connected to the alarm clock and when it goes off, the pillow will start to vibrate giving them a soft massage . A specific noise, that meets the needs of those with autism, will start to play at the same time.